backend-api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.105.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
unicorn==2.0.1.post1
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0